requirements.txt
streamlit
selenium
Pillow
streamlit-extras